Minor change to exercise 1 and 2

+++ Assignment_6_Exercise_1_Gnida.py
@@ -44,7 +44,7 @@ def RK4integrator(y0, t0 , N, f, dt):
     :return: t, Y = time grid and numerical solution
     """
 
-    y0 = np.array(y0, dtype=float)
+    y0 = np.array(y0)
 
     t = np.zeros(N + 1)
     y = np.zeros((N + 1, len(y0)))
